refactor: rename upstream_http → upstream_peer, move path resolution to ServerConfig
upstream_http was a misleading name once TLS (HTTPS) peer support was added.
upstream_peer reflects what the module actually does: construct pingora HttpPeer
values with the right TLS settings.

ServerConfig::resolve_paths() consolidates the env-var expansion, relative-path
resolution, and file-existence checks that were previously scattered as free
functions in main.rs. main.rs is now a thin orchestrator with no ad-hoc helpers.
