+++ plenum-core/src/main.rs
@@ -1,15 +1,12 @@
 use plenum_core::build_gateway;
-use plenum_core::config::{Config, ServerConfig, resolve_env_vars};
+use plenum_core::config::{Config, ServerConfig};
 
 use clap::Parser;
 
+use pingora_core::listeners::tls::TlsSettings;
 use pingora_core::server::Server;
 use pingora_core::server::configuration::{Opt, ServerConf};
 use pingora_proxy::http_proxy_service;
-
-use pingora_core::listeners::tls::TlsSettings;
-
-use std::path::Path;
 
 #[derive(Parser, Debug)]
 struct Args {
@@ -36,38 +33,6 @@ struct Args {
     openapi_overlays: Vec<String>,
 }
 
-/// Apply env var substitution (`${VAR}` / `${VAR:-default}`) to a path string.
-fn expand_env_in_path(s: String, field: &str) -> String {
-    match resolve_env_vars(serde_json::Value::String(s)) {
-        Ok(v) => v.as_str().unwrap_or("").to_string(),
-        Err(e) => {
-            eprintln!("Error in x-plenum-config {}: {}", field, e);
-            std::process::exit(1);
-        }
-    }
-}
-
-/// Resolve a path relative to config_base if it is not already absolute.
-fn resolve_config_path(s: String, config_base: &str) -> String {
-    let p = Path::new(&s);
-    if p.is_absolute() {
-        s
-    } else {
-        Path::new(config_base)
-            .join(p)
-            .to_string_lossy()
-            .into_owned()
-    }
-}
-
-/// Exit with a clear message if a required file does not exist.
-fn require_file(path: &str, label: &str) {
-    if !Path::new(path).exists() {
-        eprintln!("Error: {} not found: {}", label, path);
-        std::process::exit(1);
-    }
-}
-
 fn main() {
     env_logger::init();
     let args = Args::parse();
@@ -86,26 +51,12 @@ fn main() {
         .extension(&config.spec.extensions, "plenum-config")
         .unwrap_or_else(|_| ServerConfig::default());
 
-    // Apply env var substitution and path resolution to TLS path fields.
-    if let Some(tls) = server_config.tls.as_mut() {
-        tls.cert_path = resolve_config_path(
-            expand_env_in_path(tls.cert_path.clone(), "tls.cert_path"),
-            &args.config_path,
-        );
-        tls.key_path = resolve_config_path(
-            expand_env_in_path(tls.key_path.clone(), "tls.key_path"),
-            &args.config_path,
-        );
-        require_file(&tls.cert_path, "tls.cert_path");
-        require_file(&tls.key_path, "tls.key_path");
-    }
-    if let Some(ca_file) = server_config.ca_file.as_mut() {
-        *ca_file = resolve_config_path(
-            expand_env_in_path(ca_file.clone(), "ca_file"),
-            &args.config_path,
-        );
-        require_file(ca_file, "ca_file");
-    }
+    server_config
+        .resolve_paths(&args.config_path)
+        .unwrap_or_else(|err| {
+            eprintln!("Error in x-plenum-config: {}", err);
+            std::process::exit(1);
+        });
 
     let gateway = build_gateway(&config, &args.config_path).unwrap_or_else(|err| {
         eprintln!("Error building gateway: {}", err);
